+++ include/buzzy.h
@@ -7,9 +7,9 @@
 class buzzy
 {
 private:
-	int _pin;
-	int _noteIndex;
-	int _songLength;
+	uint8_t _pin;
+	uint16_t _noteIndex;
+	uint16_t _songLength;
 	uint8_t _loopNum;
 	uint8_t _numLoops;
 
@@ -21,17 +21,17 @@ private:
 	unsigned long _noteEndTime;
 	unsigned long _pauseEndTime;
 
-	int *_frequencies;
-	int *_durations;
-	int *_pauses;
+	uint16_t *_frequencies;
+	uint16_t *_durations;
+	uint16_t *_pauses;
 
-	void _playNote(int noteIndex);
+	void _playNote(uint16_t noteIndex);
 
 public:
-	buzzy(int pin);
-	buzzy(int pin, bool debug);
-	void playSong(int *frequencies, int *durations, int *pauses, int length);
-	void playSong(int *frequencies, int *durations, int *pauses, int length, uint8_t numLoops);
+	buzzy(uint8_t pin);
+	buzzy(uint8_t pin, bool debug);
+	void playSong(uint16_t *frequencies, uint16_t *durations, uint16_t *pauses, uint16_t length);
+	void playSong(uint16_t *frequencies, uint16_t *durations, uint16_t *pauses, uint16_t length, uint8_t numLoops);
 	void loop(void);
 	bool isPlaying();
 };
